Hide raw GitHub API helper
